requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==1.11
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.6
Pillow==5.1.0
PyJWT==1.6.1
PyMySQL==0.8.0
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
gunicorn==19.8.1
